ar71xx: allow to specify max read size for m25p80
Reading from the flash chip on the TL-WR2543ND seems buggy.
If the SPI flash driver tries to read too much data in one
SPI transfer, the flash chip returns bogus values. This can
be caused by a buggy flash chip on my board, or it can
be a bug in our SPI driver.

Add a workaround to the m25p80 driver until I find out the
root cause of the problem. The patch allows to specify the
maximum numner of bytes which can be read safely withint
one SPI transfer.

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@29679 3c298f89-4303-0410-b956-a3cf2f4a3e73